Safari-609.1.11.1/LayoutTests/http/tests/security/XFrameOptions/x-frame-options-parent-same-origin-allow.html

<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.dumpChildFramesAsText();
        testRunner.dumpResourceLoadCallbacks();
    }
</script>

<p>There should be content in the iframe below</p>
<iframe style="width:500px; height:500px" src="http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-parent-same-origin-allow.cgi"></iframe>
